+++ apisprout.go
@@ -153,7 +153,7 @@ func getTypedExample(mt *openapi3.MediaType) (interface{}, error) {
 	}
 
 	if mt.Schema != nil {
-		return OpenAPIExample(mt.Schema.Value)
+		return OpenAPIExample(ModeResponse, mt.Schema.Value)
 	}
 	// TODO: generate data from JSON schema, if no examples available?